src/app/requirements.txt
streamlit
streamlit-drawable-canvas
requests
pillow
numpy
